refactor(daemon-client): lift the daemon JSON-RPC client out of :mcp
Consuming the render-session library dragged an MCP server onto the
classpath, because `:render-session-subprocess` and
`:render-session-embedded-desktop` built their transport on `:mcp`'s
`DaemonClient`. That is the last leak the preview-server contract probe
recorded (#3824 preparation item 3).

The transport moves to a new published module, `:daemon-client`:
DaemonClient, DataProductWireException, DaemonClientFactory, DaemonSpawn,
SubprocessDaemonClientFactory, SubprocessDaemonSpawn and WorkspaceId.
Nothing MCP comes with it — no tools, no resources, no supervisor. Both
render-session modules now depend on it instead of `:mcp`, and `:mcp`
depends on the module it used to own.

The package is `ee.schimke.composeai.daemon.client` rather than the old
`ee.schimke.composeai.mcp`: two published artifacts sharing one package is
a split package, which breaks JPMS and OSGi consumers. No compatibility
shim is needed because docs/API_STABILITY.md section 1 states the
published surface of `:mcp` is its MCP tool names and input schemas — the
Kotlin types are internal and may move.

Two things deliberately did not move. DaemonLaunchDescriptorCompat stays
in `:mcp`: it is a deprecated typealias whose whole purpose is keeping an
external `import ee.schimke.composeai.mcp.DaemonLaunchDescriptor`
compiling, so relocating it would destroy the only thing it does.
DaemonClientRenderSession stays too — it is MCP-internal and nothing
outside the module references it.

SubprocessDaemonSpawnTest moves with its subject, since
SubprocessDaemonSpawn is internal and the test cannot see it from `:mcp`.
