+++ mcp/src/main/kotlin/ee/schimke/composeai/mcp/DaemonClientRenderSession.kt
@@ -1,5 +1,7 @@
 package ee.schimke.composeai.mcp
 
+import ee.schimke.composeai.daemon.client.DaemonClient
+import ee.schimke.composeai.daemon.client.DataProductWireException
 import ee.schimke.composeai.daemon.protocol.ChangeType
 import ee.schimke.composeai.daemon.protocol.DataFetchResult
 import ee.schimke.composeai.daemon.protocol.DataSubscribeResult
